fix pagination and reorganize code around credential revocations
