+++ x/credential/keeper/credential.go
@@ -5,6 +5,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
+
 	"github.com/elesto-dao/elesto/x/credential"
 )
 
@@ -24,24 +25,17 @@ func (k Keeper) GetCredentialDefinition(ctx sdk.Context, key string) (credential
 	return val.(credential.CredentialDefinition), found
 }
 
-// GetAllCredentialDefinitions
-func (k Keeper) GetCredentialDefinitions(ctx sdk.Context, req *credential.QueryCredentialDefinitionsRequest) []*credential.CredentialDefinition {
-
-	// pagition tbd
-	var cds []*credential.CredentialDefinition
-
+// GetCredentialDefinitions list credential definitions
+func (k Keeper) GetCredentialDefinitions(ctx sdk.Context, req *credential.QueryCredentialDefinitionsRequest) (cds []*credential.CredentialDefinition, pageRes *query.PageResponse, err error) {
 	store := ctx.KVStore(k.storeKey)
 	cdStore := prefix.NewStore(store, credential.CredentialDefinitionKey)
-	query.Paginate(cdStore, req.Pagination, func(key []byte, value []byte) error {
+	pageRes, err = query.Paginate(cdStore, req.Pagination, func(key []byte, value []byte) error {
 		var cd credential.CredentialDefinition
-		err := k.cdc.Unmarshal(value, &cd)
-		if err != nil {
-			panic(err)
-		}
+		k.cdc.MustUnmarshal(value, &cd)
 		cds = append(cds, &cd)
 		return nil
 	})
-	return cds
+	return
 }
 
 // SetPublicCredential persist a public verifiable credential to the store. The credential ID is used as key
