+++ x/credential/keeper/grpc_query.go
@@ -48,9 +48,8 @@ func (k Keeper) CredentialDefinitions(
 
 	ctx := sdk.UnwrapSDKContext(c)
 
-	cd := k.GetCredentialDefinitions(ctx, req)
-	return &credential.QueryCredentialDefinitionsResponse{Definitions: cd}, nil
-	// return nil, fmt.Errorf("not implemented")
+	cd, pr, err := k.GetCredentialDefinitions(ctx, req)
+	return &credential.QueryCredentialDefinitionsResponse{Definitions: cd, Pagination: pr}, err
 }
 
 func (k Keeper) PublicCredential(
